+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* TODO: its a initial release only for android to connect helix watch usingg flutter to android refer to example to know how to use.